Added level autogeneration

+++ src/main/java/com/redblueflame/herbocraft/components/LevelComponent.java
@@ -4,7 +4,7 @@ import nerdhub.cardinal.components.api.component.Component;
 
 public interface LevelComponent extends Component {
     int getHealth();
-    int getAttackSpeed();
+    float getAttackSpeed();
     int getDamage();
     int getDurability();
     int getStability();
